improved ladder climbing mechanic

+++ Assets/Scripts/Ladder.cs
@@ -5,7 +5,7 @@ public class Ladder : MonoBehaviour
 
     private void OnTriggerStay2D(Collider2D collision)
     {
-        collision.gameObject.GetComponent<Hero>().TouchingLadder((int)transform.position.x);
+        collision.gameObject.GetComponent<Hero>().TouchingLadder(transform.position.x);
     }
 
     private void OnTriggerExit2D(Collider2D collision)
